Version 1.1 - Balanced Data Set
Balanced the data set by using under-sampling on letters which appeared far more than others. Now, all entries appear the same amount of times.
*The highest amount was O's at around 56000 entries and the lowest was I's at around 1100 entries
**This means about 55000 entries for O were outright deleted!!!!

The model is actually a CONSIDERABLE amount more accurate than if we just used the full dataset, as it no longer struggles with letters such as "X", "U", "W", "M", and "N"

Final Note - The model has always struggled with the letters "Y" and "Z" since the beginning of this project, and it still does even to this iteration. The most likely solution would be to look for new datasets, possibly ones which represent these letters more.